move segment repairer to satellite
